Removed notify module
