--- src/modules/notify/notify.service.ts
@@ -1,11 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-@Injectable()
-export class NotifyService {
-    async sendMail() {
-    //   this.logger.sLog({}, 'ClientUtilsService::getEmbedding');
-      return {
-         message: 'This endpoint works'
-      };
-    }
-}
